Avoid taking error screenshot on handled errors
The espresso error handler gets invoked any time an error ocurrs when performing some action with Espresso, not after the test fails.
This means that taking a screenshot with the ErrorHandler will create screenshots for errors handled inside tests also.
With this change, we remove the error screenshot if the test method succeeds.

TL;DR: if the test passes, removes the handled error screenshot.

This fixes https://github.com/Sloy/android-screenshot-reporter/issues/4