Change the way we get the activity reference
We were using a ViewAction to wrap the action of taking a screenshot because we need the current activity. The problem with that is: if there is no activity espresso will idle until the timeout and will fail.
We're trying an alternative way now: Get the stack of activities, and obtain the first one that is Resumed. If none, we just don't take the screenshot.